tasting

+++ voh/utils.py
@@ -201,7 +201,7 @@ def pad_(ts, max_frames=None, ipad=0):
     for t in ts:
         T = t.size(-1)
         if T <= L:
-            o = torch.full((t.size(0), L), ipad, device=t.device)
+            o = torch.full((t.size(0), L), ipad, dtype=t.dtype, device=t.device)
             o[..., :T] = t
         else:
             o = t[..., :L].clone()
